moved "view" active area to base test
